requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch>=2.1
torchvision
ultralytics>=8.2
opencv-python-headless
Pillow
pyyaml
tqdm
matplotlib
seaborn
transformers
accelerate
scikit-learn